+++ app/services/payment_schedule_service.rb
@@ -158,6 +158,8 @@ class PaymentScheduleService
   end
 
   def self.cancel(payment_schedule)
+    PaymentGatewayService.new.cancel_subscription(payment_schedule)
+
     # cancel all item where state != paid
     payment_schedule.ordered_items.each do |item|
       next if item.state == 'paid'